Implement vectorial text support for `iced_tiny_skia`
